+++ NEWS.md
@@ -1,6 +1,7 @@
 # Rsrat 1.6.1
 
 * Add rateplot
+* Modify xmax of mvfplot
 
 # Rsrat 1.6.0